test: PostController test code 작성
